move SignedInUser to specific file
